pyproject.toml
[tool.poetry]
name = "django-web-shop"
version = "0.1.0"
description = ""
authors = ["stas-home <stasijanka13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.8"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
